Move auto-generatedd constraints to sub-directory
