+++ triton-vm/src/table.rs
@@ -2,18 +2,14 @@ pub mod base_matrix;
 pub mod base_table;
 pub mod challenges;
 pub mod constraint_circuit;
+pub mod constraints;
 pub mod extension_table;
 pub mod hash_table;
 pub mod instruction_table;
-pub mod instruction_table_autogen;
 pub mod jump_stack_table;
-pub mod jump_stack_table_autogen;
 pub mod op_stack_table;
-pub mod op_stack_table_autogen;
 pub mod processor_table;
-pub mod processor_table_autogen;
 pub mod program_table;
-pub mod program_table_autogen;
 pub mod ram_table;
 pub mod table_collection;
 pub mod table_column;
