+++ triton-vm/src/table/constraints.rs
@@ -1,0 +1,6 @@
+pub mod instruction_table_constraints;
+pub mod jump_stack_table_constraints;
+pub mod op_stack_table_constraints;
+pub mod processor_table_constraints;
+pub mod program_table_constraints;
+pub mod ram_table_constraints;
