+++ triton-vm/src/table/constraints/instruction_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::instruction_table::ExtInstructionTable;
+
+impl Evaluable for ExtInstructionTable {}
