+++ triton-vm/src/table/constraints/jump_stack_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::jump_stack_table::ExtJumpStackTable;
+
+impl Evaluable for ExtJumpStackTable {}
