+++ triton-vm/src/table/constraints/op_stack_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::op_stack_table::ExtOpStackTable;
+
+impl Evaluable for ExtOpStackTable {}
