+++ triton-vm/src/table/constraints/processor_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::processor_table::ExtProcessorTable;
+
+impl Evaluable for ExtProcessorTable {}
