+++ triton-vm/src/table/constraints/program_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::program_table::ExtProgramTable;
+
+impl Evaluable for ExtProgramTable {}
