+++ triton-vm/src/table/constraints/ram_table_constraints.rs
@@ -1,0 +1,4 @@
+use crate::table::extension_table::Evaluable;
+use crate::table::ram_table::ExtRamTable;
+
+impl Evaluable for ExtRamTable {}
