--- triton-vm/src/table/instruction_table_autogen.rs
@@ -1,4 +1,0 @@
-use super::extension_table::Evaluable;
-use super::instruction_table::ExtInstructionTable;
-
-impl Evaluable for ExtInstructionTable {}
