--- triton-vm/src/table/jump_stack_table_autogen.rs
@@ -1,4 +1,0 @@
-use super::extension_table::Evaluable;
-use super::jump_stack_table::ExtJumpStackTable;
-
-impl Evaluable for ExtJumpStackTable {}
