--- triton-vm/src/table/op_stack_table_autogen.rs
@@ -1,4 +1,0 @@
-use super::extension_table::Evaluable;
-use super::op_stack_table::ExtOpStackTable;
-
-impl Evaluable for ExtOpStackTable {}
