--- triton-vm/src/table/processor_table_autogen.rs
@@ -1,3 +1,0 @@
-use super::{extension_table::Evaluable, processor_table::ExtProcessorTable};
-
-impl Evaluable for ExtProcessorTable {}
