--- triton-vm/src/table/program_table_autogen.rs
@@ -1,3 +1,0 @@
-use super::{extension_table::Evaluable, program_table::ExtProgramTable};
-
-impl Evaluable for ExtProgramTable {}
