--- triton-vm/src/table/ram_table_autogen.rs
@@ -1,1 +1,0 @@
-impl Evaluable for ExtRamTable {}
